Collection

The Collection is a container of Entries, where every Entry is contains a number of samples. A sample is a part of a time series. All the samples in one Entry are linked to the same time series. These time series are containd in the SimpleFastDB (SFDB) and are linked to the Entry.

The recommended way to access time series and Layer data are through DataViews

Collection Settings

Beskriv kort och länka sen till riktiga settings The Settings contains all of the Parameters for a Collection and its Layers. Every Collection has two Settings attached to it. One, the settings_used field represent the settings that was used for the Collection. The other, the settings field, represents the settings that will be used if additional layers are added to the collection or if the layers of the colletion are redone.

Collections may also be treated as sets, with all major set functions... link to sets

Basic pruning

julia> collection["MMM"]
||Collection|| (Continuous)
Ticker: MMM
Samples: 3352

julia> collection[Date(2010,1,12)]
||Collection|| (Continuous)
Tickers: 2, MMM AOS
Samples: 2

Advanced pruning

Set operations etc, filtering etc

Entries

The Entries come in two variants, EntriesContinuous and EntriesDiscrete where the former is used for assets with a continuous time series, for example a stock at a daily resolution whereas the latter is used when the underlying time series is discontinuous, for example a stock at minute resolution (that is not traded 24/7).

EntriesContinuous is implemented as a Dict{String, Entry} where the String is some identification of the Entry. For a stock this is usually the ticker symbol. Entry

Functions

TFire.get_tickers — Method

 get_tickers(col::Collection)

Returns a vector with the names of all the tickers in the collection. For continuous entries the names are ordered as the entries.

TFire.number_of_samples — Method

number_of_samples(col::Collection)

Total number of samples for all entries in collection.

TFire.condensate_type — Method

condensate_type(col::Collection)

The type of the outermost (latest) condensate in the entries of the collection.

TFire.get_first_entry — Method

get_first_entry(collection::Collection{T}) where T <: Entries

Returns the first entry from the iterator over all the entries in the collection. Useful if one needs to get an entry but doesn't care which.

TFire.list_layer_types — Method

list_layer_types(col::Collection)

Returns a vector with the condensate types for the entries in the collection. Outermost condensate first.

TFire.layer_type — Method

layer_type(col::Collection)

Returns outermost condansate type.

TFire.copy_collection — Method

copy_collection(col::Collection)

Returns a deepcopy of the collection.

TFire.get_first_entry — Method

get_first_entry(entries::T) where T <: Entries

Returns the first entry from the iterator over all entries. Useful if one needs to get an entry but doesn't care which.

TFire.number_of_samples — Method

number_of_samples(entries::EntriesCon)

Total number of samples for all entries.

TFire.number_of_samples — Method

number_of_samples(entries::EntriesDisc)

Total number of samples for all entries.

TFire.get_tickers — Method

 get_tickers(entries::EntriesCon)

Returns a vector with the names of all the tickers in the entries. For continuous entries the names are ordered as the entries.

TFire.get_tickers — Method

 get_tickers(entries::EntriesDisc)

Returns a vector with the names of all the tickers in the entries. For continuous entries the names are ordered as the entries.

TFire.list_layer_types — Method

list_layer_types(entries::T) where T <: Entries

Returns a vector with the condensate types for the entries. Outermost condensate first.

TFire.layer_type — Method

layer_type(entries::T) where T <: Entries

Returns outermost condansate type.

TFire.get_entries_with_condansate_type — Method

get_entries_with_condansate_type(entries::EntriesCon{E}, ctype::Type{T}) where {E<:Entries, T<:Condensate}

Returns the entries with condensate type ctype. If the condensate type does not exist return nothing.

TFire.copy_entries — Method

copy_entries(entries::EntriesCon)

Returns a deepcopy of the entries.

TFire.copy_entries — Method

copy_entries(entries::EntriesDisc)

Returns a deepcopy of the entries.

TFire.delete_empty_entries! — Method

delete_empty_entries!(entries::EntriesCon)

Deletes entries with length 0 (no samples) inplace.

TFire.delete_empty_entries! — Method

delete_empty_entries!(entries::EntriesDisc)

Deletes entries with length 0 (no samples) inplace.