DataViews

EntryDataViews

DataViews are a way to access time series Layer data related to an Entry or a Collection.

For example, if e is an Entry that has a corresponding times series with the field name :volume, then dv = DVTimeSeries(e, :volume) creates a DVTimeSeries (short for DataViewTimeSeries) which is a subtype to EntryDataView. dv[i][j] then gives the j:th value of the i:th sample, or simplydv[i] gives the whole view of the volume field that is related to the i:th sample.

Likewise, if we have dvl = DVLayerField(e, :field_name, LayerType), then dv[i][j] gives j:th value of the i:th sample for the field :field_name in the Layer of type LayerType. Here, and in many other cases, the LayerType may be completely omitted if the desired Layer is the outhermost one of the Layers in the Entry.

CollectionDataViews

If instead c is a Collection then cdv = DVTimeSeries(c, :volume) creates a CDV{DVTimeSeries} (short for CollectionDataView) which contains the DVTimeSeries for every Entry in the Collection. Then cdv[1] == DVTimeSeries(e, :volume) for the first Entry e given by the iterator over all Entries. cdv[i][j][k] thus gives the k:th value of the time series of :volume corresponding to the j:th sample in the i:th Entry.

The CDV may also be flattened to a FlatCDV. If fcdv = FlatCDV(cdv) then the views for all samples in the Collection are stored in one dimension so that fcdv[i] gives the corresponding view of the :volume field for the i:th sample in the whole Collection.

The FlatCDV is useful when the value for two different fields are to be compared for the same sample without the need to know which precise sample it is. For example if the correlation between two different fields are to be calculated for all samples in a Collection.

NonRepainting DataViews

The NonRepainting Dataviews are useful when constructing NonRepainting Layers. If e is an Entry that has a corresponding times series with the field name :volume, then dv = DVNonRepaintingTimeSeries(e, :volume) creates a NonRepaintingDVTimeSeries.dv` then contains a view of the volume field from the first datetime contained in any of the i Entry's samples to the last datetime contained in any of the samples.

Functions

TFire.DVTimeSeries — Method

DVTimeSeries(entry::CollectionEntry, symbol::Symbol)

View of data from associated time series in SFDB with the name symbol.

The datetimes are matched to those in the entry so that DVTimeSeries[x][y] gives the time series value corresponding to entry[x][y]

TFire.DVNonRepaintingTimeSeries — Method

DVNonRepaintingTimeSeries(entry::CollectionEntry, symbol::Symbol)

View of data from associated time series in SFDB with the name symbol.

The datetimes are matched to those in the entry in a non repainting manner so that DVNonRepaintingTimeSeries[1:end] equals the time series entries from minimum to maximum of all datetimes inentry`. Useful for constructing non repainting layers.