Backtesting - Functions

Scoring - Functions

TFire.ScoreHold — Method

ScoreHold(score_buy::ScoreBuy, score_sell::ScoreSell, pf::PortfolioAtDate; cash_score::Float64 = 0.)

Based on a ScoreBuy and a ScoreSell, create a ScoreHold. An asset is asigned the score in ScoreBuy for all future trading days until either it is subtracted by a score in ScoreSell or it is assigned a higher score from a later date in ScoreBuy. The maximum possible score for any asset in ScoreHold is the maximum score for that asset in ScoreBuy. The minimum score is 0, i.e. score is a nonnegative number. The TradingDates are inferred from the PortfolioAtDate.

TFire.ScoreHold — Method

ScoreHold(buy_scores::ScoreBuy, sell_scores::ScoreSell, td::TradingDates; cash_score::Float64 = 0.)

Based on a ScoreBuy and a ScoreSell, create a ScoreHold. An asset is asigned the score in ScoreBuy for all future trading days until either it is subtracted by a score in ScoreSell or it is assigned a higher score from a later date in ScoreBuy. The maximum possible score for any asset in ScoreHold is the maximum score for that asset in ScoreBuy. The minimum score is 0, i.e. score is a nonnegative number.

TFire.limit_maximum_allocation — Function

limit_maximum_allocation(sh::ScoreHold, max_share::Float64=1.0)::ScoreHold

Readjusts a ScoreHold to limit the risk. max_share is the maximum share of the total score (including cash) that may be from a single asset. The removed score is transferred to cash.

TFire.score_extend — Method

score_extend(score_buy::ScoreBuy, td::TradingDates; constant_ticks::Integer=0,
linear_ticks::Integer=0, trailing_stoploss::Bool=false, trailing_stoploss_val::Float64=0.05)

TBW

Portfolio - Functions

TFire.TradingDates — Method

TradingDates(pf::PortfolioAtDate)

TBW

TFire.TradingDates — Method

TradingDates(vts::Vector{TimeSeries})

TBW

TFire.TradingDates — Method

TradingDates(col::Collection{AssetsCon{T}}) where T <: Assets

TBW

TFire.get_prices — Method

get_prices(pf::PortfolioAtDate)

TBW

TFire.get_quantities — Method

get_quantities(pf::PortfolioAtDate)

TBW

TFire.get_tickers — Method

get_tickers(pf::PortfolioAtDate)

TBW

TFire.get_total_value — Method

get_total_value(pf::PortfolioAtDate)

TBW

TFire.get_values — Method

get_values(pf::PortfolioAtDate)

Returns a vector containing the total value of each holding in the portfolio plus the cash amount.

TFire.initialize_portfolio — Method

initialize_portfolio(collection::Collection{AssetsCon{T}}, date::D, cash::Float64=1.;verbose=true) where {T <: Assets, D <: Union{Date,DateTime}}

Initializes a portfolio at a given date, usually the first trading date. Makes all assets in the collection available as holdings in the portfolio.

Portfolio Propagation - Functions

TFire.PortfolioHistory — Method

PortfolioHistory(sh::ScoreHold;start_date=DateTime(0000,01,01), start_cash=1., fee_type=:free, prop_fee=0., fixed_fee=0)

TBW

TFire.PortfolioHistory — Method

PortfolioHistory(sh::ScoreHold; start_date::D=Date(0000, 01, 01), to_date::D=Date(0000, 01, 01),
 start_cash=1.0, fee_type=:free, prop_fee=0.0, fixed_fee=0, name="") where {D<:Union{Date,DateTime}}

Construct a portfolio history by propagating a ScoreHold through time.

# Arguments
- `sh::ScoreHold`: The ScoreHold to propagate over time.
- `start_date=DateTime(0000,01,01)`: The start date for the PortfolioHistory.
- `to_date=D(0000,01,01)`: The end date for the PortfolioHistory.
- `start_cash=1.`: The starting cash amount. Default is 1.0.
- `fee_type=:free`: The fee model to use. Can be :free, :fixed, or :prop. Default is :free.
- `prop_fee=0.`: The proportional fee amount if using :prop fee type.
- `fixed_fee=0`: The fixed fee amount per trade if using :fixed fee type.

# Returns
- `PortfolioHistory`: The full portfolio history over time

TFire.propagate_portfolio — Method

propagate_portfolio(pf::PortfolioAtDate, scoring::ScoreHold; to_date::D=Date(0000,01,01), fee_type=:free, prop_fee=0.,
fixed_fee=0., keep_weighting_if_unchanged=true, name="") where D <: Union{Date, DateTime}

Propagate a portfolio forward in time by reweighting and stepping forward at each date.

# Arguments
- `pf::PortfolioAtDate`: The initial portfolio 
- `scoring::ScoreHold`: The scores to use for reweighting  
- `to_date`: The date to propagate the portfolio to. Default is the max date of the scoring.
- `fee_type`: The fee model to use. Can be :free, :fixed, :proportional.
- `prop_fee`: The proportional fee percentage if using :proportional fee model.
- `fixed_fee`: The fixed fee amount per trade if using :fixed fee model.  
- `keep_weighting_if_unchanged`: If true, keep the portfolio weighting when no change from scores.
- `name`: Name given to the propagated portfolio history.

# Returns
- `PortfolioHistory`: The propagated portfolio history.

Analysis - Functions

TFire.maximum_drawdown — Method

maximum_drawdown(pfh::PortfolioHistory)

TBW

TFire.plot_portfolio — Method

plot_portfolio(pfh::PortfolioHistory, pfh2::PortfolioHistory)

TBW

TFire.plot_portfolio — Method

plot_portfolio(pfh::PortfolioHistory)

TBW

TFire.plot_portfolio — Method

plot_portfolio(pfh_vect::Vector{PortfolioHistory})

TBW

TFire.plot_portfolio_proportions — Method

plot_portfolio_proportions(pfh::PortfolioHistory)

Plots the proportion of total portfolio value held in each asset over time.

Creates a plot with subplots for each asset showing the proportion of the total portfolio value held in that asset over time. Includes a top-level subplot showing the total portfolio value on a secondary y-axis. Limits the number of asset subplots to a reasonable amount for plot legibility.

Arguments:

pfh: PortfolioHistory containing the portfolio snapshots over time.

TFire.reference_portfolio — Method

reference_portfolio(pfh::PortfolioHistory)

Returns a portfolio with the same holdings as in pfh.

TFire.sharpe_ratio — Method

sharpe_ratio(pfh::PortfolioHistory)

TBW