Collection - Functions

TFire.condensate_type — Method

condensate_type(col::Collection)

The type of the outermost (latest) condensate in the assets of the collection.

TFire.copy_collection — Method

copy_collection(col::Collection)

Returns a deepcopy of the collection.

TFire.get_first_asset — Method

get_first_asset(collection::Collection{T}) where T <: Assets

Returns the first asset from the iterator over all the assets in the collection. Useful if one needs to get an asset but doesn't care which.

TFire.get_tickers — Method

 get_tickers(col::Collection)

Returns a vector with the names of all the tickers in the collection. For continuous assets the names are ordered as the assets.

TFire.has_layer_type — Method

has_layer_type(col::Collection, lay_typ::Type{T}) where T <: AssetLayer

Returns true if collection has layer type in chain of layers. Otherwise false.

TFire.list_layer_types — Method

list_layer_types(col::Collection)

Returns a vector with the condensate types for the assets in the collection. Outermost condensate first.

TFire.number_of_samples — Method

number_of_samples(col::Collection)

Total number of samples for all assets in collection.

TFire.setup_collection — Method

setup_collection(spec::Specification{C,T}, sfdb::SimpleFastDB) where {C<:AssetConstituents, T<:CollectionSpecification}

Returns a vector with the names of all the tickers in the collection. For continuous assets the names are ordered as the assets.

Collection Settings - Functions

TFire.add! — Method

add!(settings::Settings, param::Parameters)

Adds parameters params into settings.

TFire.add_and_replace_parameters! — Method

add_and_replace_parameters(set::Settings, set_add::Settings)

Adds all parameters in set_add to set. Replaces already existing parameters.

TFire.add_missing! — Method

add_missing!(settings::Settings, params::Parameters)

Adds parameters params into settings only if they do not already exist in setingst.

TFire.check_settings_valid — Method

check_settings_valid(settings::Settings) -> Bool

Validate Settings by checking each type should be a concrete subtype of Condensate and each parameter should be a key in the default parameters for that type, with a value of the correct type.

TFire.get_parameter — Method

get_parameter(settings::Settings, layer_type::Type, key::Symbol)

Returns the parameter value unless key is subtype of DefaultParameter in which case a special value defined by the type of key is returned.

Collection Donation - Functions

TFire.donate — Method

donate(col_reciever::Collection, col_donor::Collection;
                        dp::Union{DonationProtocol, Nothing}=nothing)

coldonor is donated to colreciever such that a new collection is created

TFire.donor_or_reciever — Method

donor_or_reciever(collection::Collection)

Returns true if collection is either a donor, a reciever or both.

TFire.dp_all_to_all — Method

dp_all_to_all(col_reciever::Collection, col_donor::Collection)

Returns a donation protocol where every ticker t in colreciever links to a vector of all tickers in coldonor.

TFire.dp_same_tickers — Method

dp_same_tickers(col::Collection)

Returns a donation protocol where every ticker t in col links to a vector of length one with the same ticker t.