External Data Handler - Functions

External Data From Disk

TFire.ExternalDataHandler.ExternalDataCollection — Method

ExternalDataCollection(edf::ExternalDataFile)

Reconstructs an ExternalDataCollection from an ExternalDataFile structure.

TFire.ExternalDataHandler.ExternalDataFile — Method

ExternalDataFile(edc::ExternalDataCollection)

Constructs an ExternalDataFile from an ExternalDataCollection. Ensures that all data assets in the collection have the same resolution, and organizes them into a dictionary of tickers for easy lookup.

TFire.ExternalDataHandler.external_data_paths_from_csv — Method

external_data_paths_from_csv(fname::String)

Given a CSV filename, extracts and returns the relative paths to data files specified in it. Ensures the input file has a CSV extension before processing.

TFire.ExternalDataHandler.fetch_external_data_from_disk — Method

fetch_external_data_from_disk(spec::SpecificationExternal, file_path::String, clean_and_prepare=true)

Retrieves external data from disk given a path. Optionally cleans and prepares the data after loading. Handles both CSV and JLD2 file formats where the CSV should be a file pointing to a list of JLD2 files.

TFire.ExternalDataHandler.load_data_from_disk — Method

load_data_from_disk(path::String, tickers::Vector{String})

Loads specific tickers' data from the provided path on the disk. Returns an ExternalDataCollection containing the loaded data.

TFire.ExternalDataHandler.load_data_from_disk — Method

load_data_from_disk(path::Vector{String}, tickers::Vector{String})

Loads specific tickers' data from multiple disk paths. Aggregates the data from all paths into a single ExternalDataCollection.

TFire.ExternalDataHandler.resolution_to_dir — Method

resolution_to_dir(resolution::String)

Maps a data resolution string (like "1d" or "1m") to its respective directory path. Helpful for organizing data storage based on the resolution of the data.

TFire.ExternalDataHandler.save_file — Method

save_file(filename::String, edf::ExternalDataFile; overwrite=false)

Persistently saves the ExternalDataFile to disk under the specified filename. If overwrite is set to true, it will replace any existing file with the same name.

TFire.ExternalDataHandler.save_to_disk — Method

save_to_disk(edc::ExternalDataCollection, file_name::String; overwrite=false)

Saves an ExternalDataCollection to the disk using the provided file name. Offers an option to overwrite an existing file with the same name.

External Data From Internet

TFire.ExternalDataHandler.fetch_external_data_from_internet — Method

fetch_external_data_from_internet(spec::Specification{C,T}, service::Type;
                    clean_and_prepare=true, verbose_error=true ) where {C<:AssetConstituents, T<:ExternalSpecification}

Fetches external data for the given constituents ext_cons using the specified service. Optionally cleans and prepares the data after fetching.

spec: A specification of which external data constituents and data fields to fetch.

service: The internet data service to use (e.g., YahooData).

clean_and_prepare: Whether or not to clean and prepare the data after fetching.

verbose_error: If true, provide detailed error messages.

Returns the fetched (and optionally cleaned) external data.