Backtesting - Functions

Scoring - Functions

TFire.ScoreHold — Method

ScoreHold(score_buy::ScoreBuy, score_sell::ScoreSell, pf::PortfolioAtDate; cash_score::Float64 = 0.)

Based on a ScoreBuy and a ScoreSell, create a ScoreHold. An asset is asigned the score in ScoreBuy for all future trading days until either it is subtracted by a score in ScoreSell or it is assigned a higher score from a later date in ScoreBuy. The maximum possible score for any asset in ScoreHold is the maximum score for that asset in ScoreBuy. The minimum score is 0, i.e. score is a nonnegative number. The TradingDates are inferred from the PortfolioAtDate.

TFire.ScoreHold — Method

ScoreHold(buy_scores::ScoreBuy, sell_scores::ScoreSell, td::TradingDates; cash_score::Float64 = 0.)

Based on a ScoreBuy and a ScoreSell, create a ScoreHold. An asset is asigned the score in ScoreBuy for all future trading days until either it is subtracted by a score in ScoreSell or it is assigned a higher score from a later date in ScoreBuy. The maximum possible score for any asset in ScoreHold is the maximum score for that asset in ScoreBuy. The minimum score is 0, i.e. score is a nonnegative number.

TFire.limit_maximum_allocation — Function

limit_maximum_allocation(sh::ScoreHold, max_share::Float64=1.0)::ScoreHold

Readjusts a ScoreHold to limit the risk. max_share is the maximum share of the total score (including cash) that may be from a single asset. The removed score is transferred to cash.

TFire.score_extend — Method

score_extend(score_buy::ScoreBuy, td::TradingDates; constant_ticks::Integer=0,
linear_ticks::Integer=0, trailing_stoploss::Bool=false, trailing_stoploss_val::Float64=0.05)

TBW

Portfolio - Functions

TFire.TradingDates — Method

TradingDates(pf::PortfolioAtDate)

TBW

TFire.TradingDates — Method

TradingDates(vts::Vector{TimeSeries})

TBW

TFire.TradingDates — Method

TradingDates(col::Collection{AssetsCon{T}}) where T <: Assets

TBW

TFire.get_prices — Method

get_prices(pf::PortfolioAtDate)

TBW

TFire.get_quantities — Method

get_quantities(pf::PortfolioAtDate)

TBW

TFire.get_tickers — Method

get_tickers(pf::PortfolioAtDate)

TBW

TFire.get_total_value — Method

get_total_value(pf::PortfolioAtDate)

TBW

TFire.get_values — Method

get_values(pf::PortfolioAtDate)

TBW

TFire.initialize_portfolio — Method

initialize_portfolio(collection::Collection{AssetsCon{T}}, date::DateTime, cash::Float64=1.) where T <: Assets

Initializes a portfolio at a given date, usually the first trading date. Makes all assets in the collection available as holdings in the portfolio.

Portfolio Propagation - Functions

TFire.PortfolioHistory — Method

PortfolioHistory(sh::ScoreHold;start_date=DateTime(0000,01,01), start_cash=1., fee_type=:free, prop_fee=0., fixed_fee=0)

TBW

TFire.propagate_portfolio — Method

propagate_portfolio(pf::PortfolioAtDate, to_date::DateTime, scoring::ScoreHold; fee_type=:free, prop_fee=0.,
fixed_fee=0., keep_weighting_if_unchanged=true, name="")

TBW

Analysis - Functions

TFire.maximum_drawdown — Method

maximum_drawdown(pfh::PortfolioHistory)

TBW

TFire.plot_portfolio — Method

plot_portfolio(pfh::PortfolioHistory, pfh2::PortfolioHistory)

TBW

TFire.plot_portfolio — Method

plot_portfolio(pfh::PortfolioHistory)

TBW

TFire.plot_portfolio — Method

plot_portfolio(pfh_vect::Vector{PortfolioHistory})

TBW

TFire.plot_portfolio_proportions — Method

plot_portfolio_proportions(pfh::PortfolioHistory)

TBW

TFire.reference_portfolio — Method

reference_portfolio(pfh::PortfolioHistory)

Returns a portfolio with the same holdings as in pfh.

TFire.sharpe_ratio — Method

sharpe_ratio(pfh::PortfolioHistory)

TBW